+++ apps/native/package.json
@@ -1,7 +1,7 @@
 {
   "name": "@cococom/mobile",
   "main": "index.ts",
-  "version": "0.1.22-alpha.cdc-18",
+  "version": "0.1.22-alpha.cdc-18.1",
   "scripts": {
     "start": "expo start --dev-client",
     "start:preview": "SUPABASE_ENV=PREVIEW expo start --dev-client",
